requirements.txt
flask == 2.0.1
werkzeug == 2.0.1 
requests >= 2.27.0
python-dotenv == 0.19.0
openai >= 1.50.0
streamlit >= 1.32.0

